refactor(connectivity): move ConnectivityStateHolder out of presentation/
DualPlayerEngine (data/service/player/) already imports it from
presentation/viewmodel/ — GEN-ARCH-01 inverted, preexisting, not
introduced by the downloads feature (R20). Moves to data/connectivity/,
a new single-file package matching existing ones (data/paging/,
data/provider/).

Moves, does not split: the file also mixes Bluetooth state (591 lines,
more than one reason to change per GEN-DES-01), but splitting it is a
refactor with its own judgment calls and doesn't belong in this
preparatory fix (F3.md §4.1).

Zero behavior change: package/import only. 9 files touched (3 resolved
it via same-package before and needed a new import; the rest just had
their import path updated). Baseline: 388 tests, same 5 pre-existing
failures, none new (check-baseline.sh). assembleDebug verified.

Known, accepted side effect: app/src/release/generated/baselineProfiles/
still reference the old presentation.viewmodel.ConnectivityStateHolder
path — those are ART hints, not code; they just stop matching for this
class until the profile is regenerated on-device, which is out of scope
here.
